Add support for the selection of child item in the table
We got new feature for the YTable, so that items in the table can have
nested items, which we were not able to select previously. This change
extends REST API functionality to accept path to the item in the table,
e.g. `root_row|sub_item_row1|sub_item_row2`.

This functionality is currently used in the expert partitioner.