Add activation for the YTable when select an item
In Qt we have issue with menu bar items not getting enabled when should
be activated by selection of the item in the table.

We need better architecture to de-couple action handling better, so for
now we just pass activation as a function, not to introduce cyclic
dependencies between YTableActionHandler and YHttpWidgetsActionHandler
classes.